requirements.txt
pillow==11.2.1
reportlab==4.4.6

